Authorization and authentication

+++ src/main/java/com/airbnb/airbnb/service/HotelServiceimpl.java
@@ -5,13 +5,16 @@ import com.airbnb.airbnb.dto.HotelInfoDto;
 import com.airbnb.airbnb.dto.RoomDto;
 import com.airbnb.airbnb.entity.Hotel;
 import com.airbnb.airbnb.entity.Room;
+import com.airbnb.airbnb.entity.User;
 import com.airbnb.airbnb.exception.ResourceNotFoundException;
+import com.airbnb.airbnb.exception.UnAuthorisedException;
 import com.airbnb.airbnb.repository.HotelRepository;
 import com.airbnb.airbnb.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,6 +35,8 @@ public class HotelServiceimpl implements HotelService{
         log.info("Creating new hotel with name"+hotelDto.getName());
         Hotel hotel=mapper.map(hotelDto,Hotel.class);
         hotel.setActive(false);
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel=hotelRepository.save(hotel);
         log.info("Created a new hotel with id{}",hotelDto.getId());
         return mapper.map(hotel,HotelDto.class);
@@ -41,6 +46,11 @@ public class HotelServiceimpl implements HotelService{
     public HotelDto gethotelbyid(Long id) {
         log.info("Getting the hotel with Id : {}"+id);
         Hotel hotel=hotelRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID"+id));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user doesn't own this hotel with id"+id);
+        }
         return mapper.map(hotel,HotelDto.class);
     }
 
@@ -49,6 +59,11 @@ public class HotelServiceimpl implements HotelService{
         log.info("Updating the hotel with Id : {}"+id);
         Hotel hotel=hotelRepository.findById(id).orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID"+id));
 
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user doesn't own this hotel with id"+id);
+        }
          Hotel savehotel=mapper.map(hotelDto,Hotel.class);
          Hotel savedhotel=hotelRepository.save(savehotel);
          return mapper.map(savedhotel, HotelDto.class);
@@ -61,6 +76,11 @@ public class HotelServiceimpl implements HotelService{
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
 
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user doesn't own this hotel with id"+id);
+        }
 
         for(Room room: hotel.getRooms()) {
             inventoryService.deleteallinventories(room);
@@ -75,6 +95,11 @@ public class HotelServiceimpl implements HotelService{
         log.info("Activating the hotel with Id : {}"+hotelId);
         Hotel hotel=hotelRepository.findById(hotelId).orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID"+hotelId));
 
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user doesn't own this hotel with id"+hotelId);
+        }
     hotel.setActive(true);
         for(Room room: hotel.getRooms()) {
             inventoryService.initializeRoomForAYear(room);
@@ -86,6 +111,12 @@ public class HotelServiceimpl implements HotelService{
         Hotel hotel=hotelRepository
                 .findById(hotelId)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID"+hotelId));
+
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user doesn't own this hotel with id"+hotelId);
+        }
         List<RoomDto> rooms=hotel.getRooms().stream().map((element)->mapper.map(element,RoomDto.class)).toList();
         return new HotelInfoDto(modelMapper.map(hotel,HotelDto.class),rooms);
    }
